agent/requirements.txt
paho-mqtt
pyyaml
numpy